+++ teneto/__init__.py
@@ -13,6 +13,7 @@ from . import generatenetwork
 from . import timeseries
 from . import misc
 from . import trajectory
+from . import temporalcommunity
 from . import communitydetection
 from .classes import TenetoBIDS, TemporalNetwork
 #del misc.teneto